TF-48 Reorganize state-management structure
- Resolve issues from review
