+++ client/src/config/constants.ts
@@ -1,0 +1,1 @@
+export const BASE_URL = 'http://localhost:5000';
