+++ client/src/store.ts
@@ -1,0 +1,10 @@
+import { createStore, applyMiddleware } from 'redux';
+import createSagaMiddleware from '@redux-saga/core';
+import { rootSaga } from './sagas';
+import { rootReducer } from './reducers';
+
+const sagaMiddleware = createSagaMiddleware();
+
+export const store = createStore(rootReducer, applyMiddleware(sagaMiddleware));
+
+sagaMiddleware.run(rootSaga);
